# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "REST API for Ludo Game with authentication and user management",
        "title": "Ludo Game API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@ludogame.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/game/{id}/move-token": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Move a token based on dice roll",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Game Actions"
                ],
                "summary": "Move a token",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Game ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Move Token Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/game.MoveTokenRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Token moved successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/game.MoveTokenResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "403": {
                        "description": "Not your turn",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "404": {
                        "description": "Game not found",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            }
        },
        "/api/game/{id}/roll-dice": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Roll the dice for your turn (returns 1-6)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Game Actions"
                ],
                "summary": "Roll the dice",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Game ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dice rolled successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/game.RollDiceResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid request or already rolled",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "403": {
                        "description": "Not your turn",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "404": {
                        "description": "Game not found",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            }
        },
        "/api/game/{id}/skip-turn": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Skip turn when no valid moves are available",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Game Actions"
                ],
                "summary": "Skip turn",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Game ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Turn skipped successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/game.SkipTurnResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid request or has valid moves",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "403": {
                        "description": "Not your turn",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "404": {
                        "description": "Game not found",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "game.MoveTokenRequest": {
            "type": "object",
            "required": [
                "steps",
                "token_index"
            ],
            "properties": {
                "steps": {
                    "type": "integer",
                    "maximum": 6,
                    "minimum": 1
                },
                "token_index": {
                    "type": "integer",
                    "maximum": 3,
                    "minimum": 0
                }
            }
        },
        "game.MoveTokenResponse": {
            "type": "object",
            "properties": {
                "coins_lost_per_player": {
                    "type": "integer"
                },
                "coins_won": {
                    "type": "integer"
                },
                "extra_turn": {
                    "type": "boolean"
                },
                "from_position": {
                    "type": "integer"
                },
                "game_completed": {
                    "type": "boolean"
                },
                "game_id": {
                    "type": "string"
                },
                "killed_opponent": {
                    "type": "boolean"
                },
                "killed_opponent_color": {
                    "type": "string"
                },
                "next_turn_color": {
                    "type": "string"
                },
                "next_turn_player_id": {
                    "type": "integer"
                },
                "player_color": {
                    "type": "string"
                },
                "player_id": {
                    "type": "integer"
                },
                "to_position": {
                    "type": "integer"
                },
                "token_index": {
                    "type": "integer"
                },
                "token_reached_home": {
                    "type": "boolean"
                },
                "winner_id": {
                    "type": "integer"
                }
            }
        },
...
        "game.RollDiceResponse": {
            "type": "object",
            "properties": {
                "can_move": {
                    "type": "boolean"
                },
                "dice_value": {
                    "type": "integer"
                },
                "extra_turn": {
                    "type": "boolean"
                },
                "game_id": {
                    "type": "string"
                },
                "player_color": {
                    "type": "string"
                },
                "player_id": {
                    "type": "integer"
                },
                "valid_tokens": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "game.SkipTurnResponse": {
            "type": "object",
            "properties": {
                "dice_value": {
                    "type": "integer"
                },
                "game_id": {
                    "type": "string"
                },
                "next_turn_color": {
                    "type": "string"
                },
                "next_turn_player_id": {
                    "type": "integer"
                },
                "skipped_player_color": {
                    "type": "string"
                },
                "skipped_player_id": {
                    "type": "integer"
                }
            }
        },
...
        "utils.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}